% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logi.reg.R
\name{bootstrap_CI}
\alias{bootstrap_CI}
\title{Bootstrap Confidence Interval}
\usage{
bootstrap_CI(x, y, alpha, n = 20)
}
\arguments{
\item{x}{a numeric vector or matrix containing explanatory variables}

\item{y}{a numeric vector containing response variable}

\item{alpha}{the designated significance level for the intervals}

\item{n}{the number of bootstraps (default is 20)}
}
\value{
a numeric matrix containing the bootstrap confidence interval
}
\description{
Calculates Bootstrap confidence intervals for the beta from \code{logreg_optim}
}
\examples{
data <- read.csv("~/Desktop/Classes/STAT 6120/Final/iris_csv.numbers")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)
x <- as.matrix(data[,1:4])
y <- data[,5]
bootstrap_CI(x = x, y, alpha = 0.05,n = 20)
}
