% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logi.reg.R
\name{conf_matrix}
\alias{conf_matrix}
\title{Confusion Matrix}
\usage{
conf_matrix(x, y, cutoff = 0.5)
}
\arguments{
\item{x}{the numeric y values of the test dataset}

\item{y}{the predicted y values of the test dataset}

\item{cutoff}{the cutoff value which by default is 0.5}
}
\value{
a list of confusion matrix and metrics such a accuracy, specificity etc.
}
\description{
confusion matrix and relevent metrics with cutoff feature.
}
\examples{
data=read.csv("~/Desktop/Classes/STAT 6120/Final/iris_csv.numbers")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)
n_features=ncol(data)
dt = sort(sample(nrow(data), nrow(data)*.7))
train<-data[dt,]
test<-data[-dt,]
x=as.matrix(train[,1:n_features-1])
y=train[,n_features]
xtest=test[,1:n_features-1]
ytest=test[,n_features]
conf_matrix(x,y)
}
