% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logi.reg.R
\name{glm_plot}
\alias{glm_plot}
\title{Logistic Regression Curve Plot}
\usage{
glm_plot(x, y, i)
}
\arguments{
\item{x}{the numeric x values of the test dataset}

\item{y}{the numeric y values of the test dataset}

\item{i}{the column of x values to be graphed}
}
\value{
a line plot of the y values and one independent variable (indicated by `i`)
}
\description{
shows a visual of logistic regression values against a variable.
}
\examples{
data <- read.csv("~/Desktop/Classes/STAT 6120/Final/iris_csv.numbers")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)
x <- as.matrix(data[,1:4])
y <- as.vector(data[,5])
glm<- glm_plot(x,y,2)
}
