% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logi.reg.R
\name{logreg_optim}
\alias{logreg_optim}
\title{Logistic Regression Using Numerical Optimization}
\usage{
logreg_optim(x, y, alpha, n = 20)
}
\arguments{
\item{x}{a numeric vector or matrix containing explanatory variables}

\item{y}{a numeric vector containing response variable}
}
\value{
a \code{vector} that contains estimated coefficients of the logistic
regression model
}
\description{
Estimates a beta coefficient from a set of explanatory and response variables
with basic logistic regression using numerical optimization. Also displays initial coefficients used for
the regression from the least squares formula.
}
\examples{
data <- read.csv("/Users/ebyrne1/Desktop/iris_csv.csv")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
logreg_optim(x, y)
}
