% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logi.reg.R
\name{metricplot}
\alias{metricplot}
\title{Metric Plot}
\usage{
metricplot(x, y, metric)
}
\arguments{
\item{x}{the numeric y values of the test dataset}

\item{y}{the predicted y values of the test dataset}

\item{metric}{is the name of mr=etric for which the plot is required, like Accuracy.}
}
\value{
any required metric plot against a range of cutoff values from 0.1-0.9
}
\description{
provides any required metric plot against a range of cutoff values.
}
\examples{
data <- read.csv("~/Desktop/Classes/STAT 6120/Final/iris_csv.numbers")
data=data[data$class=="Iris-setosa" | data$class=="Iris-virginica",]
data$class=ifelse(data$class=="Iris-setosa",1,0)
x <- as.matrix(data[,1:4])
y <- as.vector(data[,5])
metricplot(x,y,metric="Accuracy")
}
